Refactor: update component's file name to be consistent -> from overView to over-view
